moved modals for directory clarity
