+++ js/admin/src/modals/BazaarSettingsModal.js
