+++ js/admin/src/modals/FilePermissionsModal.js
